fix: confused naming and types for Zielnorm references
RISDEV-7625
